requirements.txt
africastalking==1.1.4
astroid==2.0.4
autopep8==1.4.1
certifi==2018.10.15
chardet==3.0.4
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==1.11
django-bootstrap4==0.0.7
django-braces==1.13.0
django-heroku==0.3.1
django-model-utils==3.1.2
django-registration==2.4.1
gunicorn==19.9.0
idna==2.7
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
Pillow==5.3.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
pycodestyle==2.4.0
pylint==2.1.1
python-decouple==3.1
pytz==2018.5
qrcode==6.0
requests==2.20.0
schema==0.6.8
six==1.11.0
swapper==1.1.0
typed-ast==1.1.0
urllib3==1.24
whitenoise==4.1
wrapt==1.10.11
